% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinearRegression.R
\name{plotLinearRegression}
\alias{plotLinearRegression}
\title{A function to plot a linear regression.}
\usage{
plotLinearRegression(x, y, group = NULL, dataset = rec_rate)
}
\arguments{
\item{x}{numeric. Continuous variable.}

\item{y}{numeric. Continuous variable}

\item{group}{character. Defines a grouping variable.}

\item{dataset}{data.frame. Default is rec_rate.}
}
\description{
A function to plot a linear regression.
}
