% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLinearRegression.R
\name{runLinearRegression}
\alias{runLinearRegression}
\title{A function to run a linear regression.}
\usage{
runLinearRegression(x, y, group = NULL, dataset = rec_rate)
}
\arguments{
\item{x}{numeric. Continuous variable.}

\item{y}{numeric. Continuous variable}

\item{group}{character. Defines a grouping variable.}

\item{dataset}{data.frame. Default is rec_rate.}
}
\description{
A function to run a linear regression.
}
